Created TeleOp
Created a teleop mode that allows for us to drive the robot as well as moving the clamp up and down while also closing and opening it. We also allow for control of the piston which allows for us to launch blocks onto the balancing pads during autonomous. Created a baseline autonomous that drives until the robot gets within 30 centimeters of something and then spins.